requirements.txt
PyYAML
asgiref==3.2.7
asn1crypto==1.3.0
attrs==19.3.0
beautifulsoup4==4.9.0
bleach==3.1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap4==1.1.1
django-cors-middleware==1.5.0
django-filter==2.2.0
django-heroku==0.3.1
django-markdownify==0.8.0
django-markup==1.4.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.16
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itypes==1.1.0
Jinja2==2.11.2
Markdown==3.2.1
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.4
psycopg2-binary==2.8.5
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==5.4.1
pytest-django==3.9.0
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==5.0.1
zipp==3.1.0
